+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 1. Enable poap scheduling (POAP API integration)
 2. Remove Datadog
 3. Enable poap config via channel flow
+4. Add poap config status command to display list of authorized users and roles
+5. Rename `/poap schedule` -> `/poap mint`
 
 ## 2.2.0-RELEASE (2021-11-30)